Split git testing utilities and performance testing into separate files
